requirements.txt
django>=4.2,<6.1
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
Jinja2>=3.1
Pillow>=10.0
